requirements.txt
requests==2.32.4
beautifulsoup4==4.12.2
python-telegram-bot==20.5
